Big refactor for my mental sanity
