+++ package.json
@@ -14,6 +14,7 @@
     "@editorjs/editorjs": "^2.28.2",
     "@editorjs/header": "^2.7.0",
     "@editorjs/paragraph": "^2.11.2",
+    "@types/node": "^20.8.8",
     "axios": "^1.5.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
