+++ src/App.tsx
@@ -1,27 +1,11 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { UserContextProvider } from "./store/userContext";
-import Homepage from "src/pages/homepage/Homepage";
-import Articles from "./pages/articles/Articles";
-import Navbar from "./auth/Navbar";
-import ArticleEditor from "./pages/articles/ArticlesEditor";
-import ArticleView from "./pages/articles/ArticleView";
-
-import { useEffect } from "react";
-import { useLocation } from "react-router-dom";
-
-export function ScrollToTop() {
-  const { pathname } = useLocation();
-
-  useEffect(() => {
-    window.scrollTo({
-      top: 0,
-      left: 0,
-      behavior: "smooth",
-    });
-  }, [pathname]);
-
-  return null;
-}
+import { UserContextProvider } from "./utils/userContext";
+import Homepage from "src/pages/home/Homepage";
+import Articles from "./pages/articles/home/Articles";
+import Navbar from "./components/Navbar";
+import ArticleEditor from "./pages/articles/editor/ArticleEditor";
+import ArticleView from "./pages/articles/view/ArticleView";
+import ScrollToTop from "src/utils/scrollToTop";
 
 export default function App() {
   return (
