+++ src/index.css
@@ -29,8 +29,49 @@ h6.ce-header {
 
 .ce-block__content,
 .ce-toolbar__content {
-  max-width: calc(100% - 80px) !important;
+  max-width: calc(100% - 120px) !important;
 }
 .cdx-block {
   max-width: 100% !important;
+}
+
+/* Set the background color of the Editor.js container element */
+.ce-block {
+  color: #f5f5f5;
+}
+
+/* Set the background color of the Editor.js toolbar */
+.ce-toolbar__content {
+  background-color: #2d2d2d;
+}
+
+/* Set the color of the toolbar plus icon */
+.ce-toolbar__plus {
+  color: #f5f5f5;
+}
+
+/* Set the color of links in the Editor.js content blocks */
+.ce-block__content a {
+  color: #f5f5f5;
+}
+
+/* Set the color of links in the Editor.js content blocks when hovered */
+.ce-block__content a:hover {
+  color: #f5f5f5;
+  text-decoration: underline;
+}
+.ce-toolbar__settings-btn svg {
+  color: #f5f5f5 !important;
+}
+.ce-inline-toolbar__toggler-and-button-wrapper {
+  color: #000 !important;
+}
+.ce-conversion-toolbar__tools {
+  color: black;
+}
+.codex-editor ::selection {
+  background-color: theme("colors.highlight") !important;
+}
+.ce-block--selected .ce-block__content {
+  background-color: theme("colors.highlight") !important;
 }
